requirements.txt
aiohttp
botbuilder-core
botbuilder-schema
botbuilder-integration-aiohttp